requirements.prod.txt
fastapi
sqlalchemy
uvicorn
pymysql
python-dotenv
rich
gunicorn
openai==1.55.3
google-cloud-bigquery
google-cloud-vision
google-cloud-aiplatform
google-cloud-storage
loguru
pillow
db-dtypes
pydantic_settings
azure-cognitiveservices-vision-customvision
azure-storage-blob
chromadb